distribute URLs by host
